% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clinical_timeline_block}
\alias{clinical_timeline_block}
\title{Clinical timeline block}
\usage{
clinical_timeline_block(data, ...)
}
\arguments{
\item{data}{Dataset.}

\item{...}{Ignored.}
}
\description{
Clinical timeline block
}
