% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{new_clinical_timeline_block}
\alias{new_clinical_timeline_block}
\title{Clinical timeline block}
\usage{
new_clinical_timeline_block(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Clinical timeline block
}
