% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{new_clinical_timeline_data_block}
\alias{new_clinical_timeline_data_block}
\title{Clinical Timeline Data Block}
\usage{
new_clinical_timeline_data_block(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
This data block is used to create a clinical timeline data frame from the ADaM data sets.
}
